+++ src/just_makeit/templates/c/src/component_ext.c
@@ -6,7 +6,7 @@
 #include <Python.h>
 #define NPY_NO_DEPRECATED_API NPY_1_7_API_VERSION
 #include <numpy/arrayobject.h>
-#include <complex.h>
+#include "clib_common.h"
 
 #include "/*<<component>>*///*<<component>>*/_core.h"